+++ shaders/effects/classicNoisedeck/mediaPost/definition.js
@@ -83,7 +83,7 @@ export default new Effect({
       }
     },
     rotation: {
-      type: "int",
+      type: "float",
       default: 0,
       uniform: "rotation",
       min: -180,
