+++ shaders/effects/classicNoisedeck/tunnel/definition.js
@@ -63,7 +63,7 @@ export default new Effect({
       }
     },
     speed: {
-      type: "int",
+      type: "float",
       default: 1,
       uniform: "speed",
       min: -5,
@@ -75,7 +75,7 @@ export default new Effect({
       }
     },
     rotation: {
-      type: "int",
+      type: "float",
       default: 0,
       uniform: "rotation",
       min: -5,
